Made Header and Footer read context objects.
