+++ client/src/components/Footer.tsx
@@ -3,19 +3,16 @@ import GoogleMap from './GoogleMap';
 import './Footer.css';
 import Toe from './Toe';
 import { LocaleContext, locales } from '../contexts';
-
-type StringDict = {
-    [key: string]: string
-};
+import { ContentDict, ContentObject, ContentText } from '../types/api_object_types';
 
 type Props = {
-    contentSv: StringDict,
-    contentEn: StringDict
+    contentSv: ContentObject,
+    contentEn: ContentObject
 }
 
 function Footer(props: Props) {
     const locale = useContext(LocaleContext);
-    const content = locale === locales.sv ? props.contentSv : props.contentEn;
+    const content = (locale === locales.sv ? props.contentSv : props.contentEn) as ContentDict;
 
     return (
         <div>
@@ -23,14 +20,14 @@ function Footer(props: Props) {
 
                 <h3>{locale === locales.sv ? 'Hitta hit!' : 'Find us!'}</h3>
 
-                {content.address}
+                {(content.items.address as ContentText).text}
 
                 <GoogleMap />
 
             </div>
             <Toe
-                webmaster={content.webmaster!}
-                currYear={content.currYear!}
+                webmaster={(content.items.webmaster as ContentText).text}
+                currYear={(content.items.currYear as ContentText).text}
             />
         </div>
     );
