+++ client/src/types/general.ts
@@ -3,7 +3,7 @@
  */
 
 import React from 'react';
-import { Page } from './api_responses';
+import { Page } from './api_object_types';
 
 export type PageComponentProps = {
     data: Page
